Render highway=primary+expressway=yes in its own style.
As described in #22, `highway=primary + expressway=yes` is rendered as 
as heavy casing with a light fill to contrast with non-expressway 
(solid) primary roads. The fill width has also been narrowed slightly to 
make the casing stroke perceived as heavier and more significant 
compared to the normal think black casing of non-expressway roads.