Render Expressways in their own style (#82)
* Render highway=trunk+expressway=yes in its own style.

As described in #22, `highway=trunk + expressway=yes` is rendered as as 
heavy casing with a light (almost white) fill to contrast with 
non-expressway (solid) trunk roads. The fill width has also been 
narrowed slightly to make the casing stroke perceived as heavier and 
significant compared to the normal think black casing of non-expressway 
roads.

Overall casing & fill widths have been chosen to present a line that has 
a visual weight halfway between that of motorways and other trunks at 
most zooms.

* Render highway=primary+expressway=yes in its own style.

As described in #22, `highway=primary + expressway=yes` is rendered as 
as heavy casing with a light fill to contrast with non-expressway 
(solid) primary roads. The fill width has also been narrowed slightly to 
make the casing stroke perceived as heavier and more significant 
compared to the normal think black casing of non-expressway roads.

* Render highway=secondary+expressway=yes in its own style.

As described in #22, `highway=secondary + expressway=yes` is rendered 
as heavy casing with a light fill to contrast with non-expressway 
(solid) primary roads. The fill width has also been narrowed slightly to 
make the casing stroke perceived as heavier and more significant 
compared to the normal think black casing of non-expressway roads.

* Render highway=tertiary+expressway=yes in its own style.

As described in #22, `highway=tertiary + expressway=yes` is rendered 
as heavy casing with a light fill to contrast with non-expressway 
(solid) primary roads. The fill width has also been narrowed slightly to 
make the casing stroke perceived as heavier and more significant 
compared to the normal think black casing of non-expressway roads.

* Add expressway width level to allow tweaking across the zoom range.

* Make expressway fills match other roads and casings extend outboard.