Load pre-built custom mapbox-gl from a Gist, include in bundled artifact
Using a Gist as a temporary hosting mechanism. Based on the [npm 
documentation](https://docs.npmjs.com/cli/v8/commands/npm-install) it 
looks like `npm install <tarball url>` might be a better option for 
hosting the maplibre-gl fork, but that should be built at 
https://github.com/ZeLonewolf/maplibre-gl-js and include the 
`package.json` in the current Gist.