Force forked maplibre-gl to be used by dependencies. Requires npm 8.3+.
openmapsamples-maplibre also needs maplibre-gl and without this override
was getting a newer version installed, which was then conflicting with
our forked version that is a direct dependency. This override feature is
new as of npm v8.3, so the `engines` entry and `.npmrc` are added to
prevent install issues due to older npm ignoring the override.