Use "NSP" as the ref for Niagara Scenic Parkway.
Currently, the road is tagged as being part of the US:NY:Parkway
network, due to the similarity between its shield and other green
parkway shields. While there still are differences, including the first
letter of each word not being emphasized, the shield should still
probably be rendered the same way as others in the network.

Should the road be retagged under its own network, this should be
changed.