Workspace tweaks (#962)
* Don't depend on a file: version.

Since shieldlib is listed in workspaces, it will be installed into
node_modules as a dependency, assuming it is semver compatible (by
depending on *, it will always be).

The main americana project could also be a workspace, but that would
necessitate breaking some URLs. It's fine for the root project to list
the workspaces and itself be a buildable artifact, but pushing it down
to a subdirectory might be desirable.

This is the sort of thing that a precommit/prepublish hook should
enforce, maybe.

As a "file:" dependency, both pnpm and yarn copy the directory into
node_modules/, which is undesired as it will not have been built yet.

* Remove double dependency on glob.

* Pacify typedoc dependency

* Remove npx, exec, and npm exec from scripts

exec is redundant but is too sophisticated for yarn's shell
implementation. npx and npm exec are redundant and additionally allow
wandering outside the project to find globally-installed packages.

* Hoist typedoc

This is a concession to yarn not finding bin's from workspaces

* Pacify ts-mocha with a peer dependency

* Unhoist dependency

Instead of a script calling 'npm run', use npm-run-all, which is clever
enough to spawn a yarn/pnpm process.