swap out ts-node for tsx
apparently node 18.19.0 decided to change the (admittedly experimental)
import hooks in a backwards-incompatible way. ts-node has not had a
release to support this new interface, so switch to tsx which has.

incidentally this also means the scripts are now compiled with esbuild
instead of with typescript, but we already depend on esbuild.
(and it's been a TODO to actually run tsc to typecheck everything.)