Add initialism for Bear Mountain Parkway
The name in OSM was changed to remove the "State" from "Bear Mountain
State Parkway", since it is usually omitted in speech and is not signed
on the ground at all. However, since it is possible that someone might
change it back, this commit does not remove the mapping for the old
name.