[a11y] Search box screenreader support
Use specific tags and roles: The search input is a combobox, the results
are a list.

When selecting the search results, use the ariaActiveDescendantElement
property to cause the screen reader to highlight/announce the item. This
is a very new property, but it should gracefully degrade as we are only
assigning to it.